backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.0",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.12",
    "uvicorn[standard]==0.32.0",
    "livekit-agents~=1.2",
    "livekit-plugins-openai~=1.2",
    "livekit-plugins-silero~=1.2",
    "livekit-plugins-cartesia~=1.2",
    "livekit-plugins-deepgram~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "certifi>=2025.10.5",
]
